pyproject.toml
[project]
name = "agentic-ai"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "ipykernel>=6.30.1",
    "jupyter>=1.1.1",
    "jupyterlab>=4.4.7",
    "openai>=1.107.0",
    "pdfplumber>=0.11.7",
    "pypdf>=6.0.0",
    "python-dotenv>=1.1.1",
    "rich>=14.1.0",
]

[tool.uv.workspace]
members = [
    "site-summarizer",
]
